test: Added tests for setAdditionalRecipeInformation method
